Splitting app into three variants
